// src/types/partials/HealthcareServiceAvailableTime.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TTime } from '../simpleTypes/Time';

export type THealthcareServiceAvailableTime = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    daysOfWeek?: String[];
    allDay?: Boolean;
    availableStartTime?: TTime;
    availableEndTime?: TTime;
};